backend/pyproject.toml
[project]
name = "superconducting-circuits-backend"
version = "0.1.0"
description = "FastAPI rewrite foundation for superconducting circuits services."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.0",
    "pydantic-settings>=2.11.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.34.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.14.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "UP",
    "N",
    "SIM",
    "RUF",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.basedpyright]
typeCheckingMode = "basic"
pythonVersion = "3.12"
include = ["src", "tests"]
reportMissingTypeStubs = "none"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
